Change the type of values that we get from the API

+++ app/src/main/java/com/example/lomakincountriesapp/data/Country.kt
@@ -4,7 +4,7 @@ data class Country(
     var name: Name?,
     var capital: ArrayList<String>,
     var languages: Map<String?, String?>,
-    var area: Double,
-    var population: Int,
+    var area: Long,
+    var population: Long,
     var flags: Flags?,
 )
